Refactor: move BulkEditButton for consistency
